Now supports GNU style long filenames.
If no filenames are specifed to be (extracted/listed) it will act on
all files in the archive (i dont think the old version did this)

Code is structured to better support upcoming recursive extraction
functionality.

The old code was a bit strange in places, dont know what i was thinking.


git-svn-id: svn://busybox.net/trunk/busybox@1018 69ca8d6d-28ef-0310-b511-8ec308f3f277