This is the initial checkin of Larry Doolittle's hush.c, using his April 25, 2001 snapshot, adjusted a bit by me so it has cmdedit support.
This checkin also removes sh.c.  In the future sh.c will be a symlink to
your shell of choice.  For now, this symlink will default to pointing to
lash.c (as in the past).  If you change the symlink to point to hush.c,
then thats what you will get.  This symlink business is a temporary situation,
which will be cleaned up Real Soon Now(tm).

 -Erik


git-svn-id: svn://busybox.net/trunk/busybox@2466 69ca8d6d-28ef-0310-b511-8ec308f3f277