Stephane Billiart writes:
bb_lookup_port now takes 3 parameters but rdate has not been modified
accordingly and fails to compile in the current CVS version.
The modification below fixes the problem.

Now, RFC868 allows both UDP and TCP implementations of the time protocol
so this may not work if someone defines a udp time service other than 37
but who would do that?


git-svn-id: svn://busybox.net/trunk/busybox@8314 69ca8d6d-28ef-0310-b511-8ec308f3f277