Rob Landley writes:
While building glibc with busybox as part of the development environment, I
found a bug in glibc's regexec can throw sed into an endless loop.  This
fixes it.  Should I put an #ifdef around it or something?  (Note, this patch
also contains the "this is not gnu sed 4.0" hack I posted earlier, which is
also needed to build glibc...)


git-svn-id: svn://busybox.net/trunk/busybox@8398 69ca8d6d-28ef-0310-b511-8ec308f3f277