Richard Kojedzinszky writes:
Hi All,

I aplogoize for the mistake, but i have just recognized that somehow the
last patch I sent in was wrong, and a '0' was instead of a '-1'. Because
of this, vi does behave the wrong way. So again, it should be the last
patch for vi. This is for pre7.


git-svn-id: svn://busybox.net/trunk/busybox@8422 69ca8d6d-28ef-0310-b511-8ec308f3f277