Togg writes:
Syslogd wont start if remote-logging is enabled and the connection to the
remote-log server is not possible on syslogd startup.

I found a patch somewhere which works like a charm. It uses sendto() which
seems more reliable for this issue.

Please see attached patch. Many people will be more happy with this included
I think.

Regards,
Togg


git-svn-id: svn://busybox.net/trunk/busybox@9171 69ca8d6d-28ef-0310-b511-8ec308f3f277