Takeharu KATO writes:
Hi,

I found that gcc in cvs (HEAD in 2005/02/11) reject the gzip source
in the busybox.

This is caused by changing gcc's error handling behavior(
The gcc check the function prototype more strictly).

I show the compilation log as follow:
-- compilation log

-- compilation log
To fix the problem, apply the patch which is attached with this
mail.

Please take a look the patch and apply the patch into svn repository.




git-svn-id: svn://busybox.net/trunk/busybox@9849 69ca8d6d-28ef-0310-b511-8ec308f3f277