This one's from me.  Fix ash "standalone shell".
If we exec /proc/self/exe and only fall back to /bin/busybox if /proc isn't
there, then we have a reasonable chance of having the standalone shell work
even if busybox isn't installed in /bin on the system in question.

Still won't work in a chroot environment, but it's an improvement.



git-svn-id: svn://busybox.net/trunk/busybox@10264 69ca8d6d-28ef-0310-b511-8ec308f3f277