Fix the warnings, and fix the following two obvious segfaults: 	./busybox getopt -n one -n two woot 	./busybox getopt -o one -o two woot
This entire applet is still an enormous pile of garbage, which I can't clean
up because I really have no idea what it's for.  (Both "man getopt" and trying
it out on the command line a bit fail to enlighten me.  Reading the code, the
fact half of it seems to be special cases for bash vs tcsh does not fill me
with confidence.)


git-svn-id: svn://busybox.net/trunk/busybox@11326 69ca8d6d-28ef-0310-b511-8ec308f3f277