Don't comment warnings.  Don't put #warnings about warnings.  _FIX_ warnings.
This fixes the warning, and makes the binary smaller out of sheer pique.
(Yes, since Manuel did this one it's nice tight code that took several
attempts to shrink, but I was ticked.)

Add the start of a test for uniq; this is about the first 1/3 of the
tests we need for full susv3 coverage of uniq.


git-svn-id: svn://busybox.net/trunk/busybox@11347 69ca8d6d-28ef-0310-b511-8ec308f3f277