some combinations of status line and screen refresh don't give a correct screen, and bug 215 reports trouble with the status line on small screens.
with this change a) the status line should always be refreshed
properly, b) the status line is a little shorter than it used to
be ("I" instead of "--INSERT--"), c) the status line will be
truncated if it doesn't fit on the screen, and d) if the screen
is too narrow for an error or transient status message (from
psb() or psbs()), then that message will be followed by a "Hit
Return" prompt.  (it wasn't until i did this last bit that the
size grew.  with this, these changes add about 150 bytes.)

- pgf



git-svn-id: svn://busybox.net/trunk/busybox@11477 69ca8d6d-28ef-0310-b511-8ec308f3f277