fix bug #474:      0000474: vi crashes often
problem was that the buffer used for "." command ("last_modifying_cmd")
wasn't being maintined correctly -- the recording code was walking back
over the front of that buffer when a repeatable insert command
included backspacing (e.g. "i\b\b\bfoo").  the fix is to simply
record the backspaces along with the rest of the command.
    
also, cleaned up start_new_cmd_q() slightly.



git-svn-id: svn://busybox.net/trunk/busybox@12559 69ca8d6d-28ef-0310-b511-8ec308f3f277