Roy Walker writes:
Here is a patch against the current subversion repository, that makes
udhcpc have an adjustable timeout.  Works for both foreground and before
it drops to the background.  This brings it more in-line with ISC dhcpc.

Use like so:
udhcpc --timeout=10 ...
or
udhcpc -T 10 ...

Still shooting for 1.1.1 this month?  Would really be great if you could
get this in that release.

Please give credit to Paul Pacheco - ppacheco@gmail.com.



git-svn-id: svn://busybox.net/trunk/busybox@14581 69ca8d6d-28ef-0310-b511-8ec308f3f277