I touched perl.  I feel dirty.
Make autodocifier suck less.  It still doesn't handle nested USE( USE() ) case
(the inner USE() winds up in the output), but making it recursive involves
getting perl to accept a "for" loop and it's telling me that "break" is an
unrecognized bareword and I hate perl.  This is at least an improvement.


git-svn-id: svn://busybox.net/trunk/busybox@15751 69ca8d6d-28ef-0310-b511-8ec308f3f277