So, in the cornucopia of superfulous warning directives I didn't add to the tree, we have the warning about failed inlines (which with our inline limit set to zero means any usage of the "inline" keyword at all).  Note that setting the inline limit to zero, and using -Werror, both predated adding the warning about inlines to the tree.  So whatever checkin added that did nothing but break the tree.  But oh well.
The second category of superfluous warnings is warning about functions with
no previous declaration.  Apparently, if you add ALWAYS_INLINE to an empty
function definition, it considers the sucker undeclared as far as the
warning is concerned.  (I.E. it's a buggy warning.  I try not to ask the
compiler to generate warnings it can't competently generate.)

This is why I removed "inline" (unbreak allbareconfig), and couldn't replace it
with "ALWAYS_INLINE" (still broke allbareconfig).


git-svn-id: svn://busybox.net/trunk/busybox@16126 69ca8d6d-28ef-0310-b511-8ec308f3f277