ash: reset pointers to builtin environment variables after fork
The builtin environment variables can be accessed using macros
that reference the varinit array.  initvar puts pointers to varinit
into the variable hash table.  During forkshell_prepare two copies
of the builtin variables are made:  once as the varinit array and
again through the pointers in the hash table.  One of these copies
is accessed by code that uses the macros and the other by code that
looks up the variable by name.

This is the cause of the strange behaviour of IFS in backticks:

   https://github.com/pclouds/busybox-w32/issues/12

To avoid the problem the pointers in the hash table are reset to
the varinit array in forkshell_init.  It seemed easier to do it
this way than to try and prevent the duplicate copies being made
in the first place.