Use putenv to implement unsetenv/clearenv
noexec applets failed on ReactOS 0.3.17.  This was because the
environment was being manipulated directly using the environ pointer.
Implementing unsetenv and clearenv using putenv fixes the problem.

WIN32 putenv doesn't allow environment variables to have empty
values.  This was the case before and it's still the case after
this change.  Shell variables are fine.