ar: fix modification of existing archives
When modifying an existing archive the code opens the old file, unlinks
it and creates a new file with the same name.  This doesn't work on
Windows where it isn't possible to unlink an open file.  Instead we
create a temporary file for output and change it's name when we're done.