test: fix access check bug; tweak fake file mode
Changing the fake uid from 0 to 1000 uncovered a problem with file
access checks in test.  Previously the tests were short-circuited
because it thought we were root.  With a non-root user the tests
were performed, but the mode bits had been redefined for WIN32.

Also adjust the fake file mode returned by stat so that the user
and group modes are identical.  Other users have the same modes
as user but without write permission.