ash: skip poll in read built-in
GNUlib poll doesn't seem to work when the file descriptor refers
to a file on disk.  This may not be surprising given the warning
in the Windows API documentation against using disk file handles
in calls to WaitForSingleObject.

Skipping poll in the read built-in allows file redirection to
work and doesn't make interrupt handling any worse than it was
before.