Enable seamless compression for WIN32
In the archival code we pretend that WIN32 is a no-MMU platform and
use the new mingw_popen_fd routine to pipe data to/from commands
to compress/decompress.

The pretence is maintained by redefining MMU macros in bb_archive.h.
This is mostly used in the archival code but there are a handful
of places where it's used to access public interfaces.  The symbol
BB_ARCHIVE_PUBLIC is defined in these places.

With these changes:

   tar supports seamless compression/decompression
   rpm2cpio and dpkg-deb can be enabled