ash: add winxp option
In some circumstances Windows XP and Windows Server 2003 seem to dislike
environment variables with backslashes in paths.

The 'winxp' option prevents BusyBox from changing backslashes to slashes
in the environment.  Although the option can be managed by the 'set'
builtin it only works if '-X' is given as the first argument to the shell.