Allow different ways to manipulate environment for XP vs 64-bit
Commit fa147bd 'Use putenv to implement unsetenv/clearenv' allowed
BusyBox to run on ReactOS but broke it on Windows XP so it was
reverted.

It turns out that the same change is required on 64-bit Windows.
Reinstate the 'safe' environment manipulation code but make it a
configuration option.  Add a config file for 64-bit Windows that
does the right thing.