ash: adjustment to 'busybox' command in standalone shell mode
A previous change allowed the command 'busybox' (which isn't an
applet) to be invoked by a direct call to the executable, avoiding
a PATH lookup.

For some reason I also checked that the name of the executable was
'busybox.exe'.  This seems unnecessary and prevents the trick from
working when the binary is called 'sh.exe', so don't do it any more.