winansi: speed up clearing of screen
Various tasks on the shell command line (e.g. backspace, DEL,
reverse search) were found to be slow when the console screen
buffer was very large.  This was because the entire buffer was
being cleared in erase_till_end_of_screen().

Rewrite erase_till_end_of_screen() to only clear to the end of the
visible screen and to be tidier.  (Based on a suggestion by GitHub
user avih.)

Also, modify move_cursor to use coordinates relative to the current
display (as ANSI escapes expect) rather than relative to the screen
buffer.