winansi: revert to previous console behaviour for vi/less
Recent changes to make the Windows console behave more like a *nix
terminal didn't work too well for vi/less.  On *nix the terminal
buffer can't be scrolled while such screen-based applications are
running.  In the Windows console this remained possible and led to
confusion.

Add a new routine to allow vi/less to revert to their previous
behaviour where the cursor is positioned at the top of the buffer
and the entire buffer is cleared.