Use %I64* formats with MinGW instead of %ll* formats
The MSVC runtime uses the format specified %I64 for 64-bit data types;
It does not understand e.g. %llu for unsigned long longs.

However, mingw-w64 provides a compatibility mode in its runtime wrapper
which can be activated by defining the constant __USE_MINGW_ANSI_STDIO=1
in which case we must refrain from overriding the %ll* formats.

This fixes quite a couple of compile warnings when building with the
mingw-w64 compiler.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>