kbuild: simulate force-renaming on Windows
Calling rename() when a file with the target name exists already fails
on Windows.

Let's simply delete the target file, if it exists. This is usually a
dangerous thing as it leaves time between deleting and renaming, during
which (theoretically) another thread or process could have created the
same file, and the rename() would fail again. Practically, we are
talking about the kbuild system used to configure the build, where it
does not matter, as we do not expect any other thread or process to
interfere with the configuration files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>