xconnect: avoid using getsockname/getpeername on Windows
The semantics of said functions is a bit different on Windows than on
Linux: it takes a `SOCKET` parameter instead of an `int` parameter, as
Windows does not use file descriptors for sockets.

The callers of these functions are only used in the daemons anyway,
though, and those are not compiled in busybox-w32.

Therefore, let's simply skip compiling those functions on Windows (as
well as `get_lsa()`, which would now be unused otherwise).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Ron Yorston <rmy@pobox.com>