mingw: accommodate for BusyBox' assumptions about isatty()
On Windows, isatty(fd) determines whether the file descriptor refers to
a character device.

The thing is: even NUL or a printer is a character device. BusyBox
thinks, however, that isatty() only returns non-zero for an interactive
terminal.

So let's shadow isatty() by a version that answers the question BusyBox
wants to have answered.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Ron Yorston <rmy@pobox.com>