ash: improve performance of ctrl-c
The WIN32 port uses a console control handler to detect crtl-c.  This
replaces the signal handler used in POSIX.  The control handler sets
an event that can be detected in waitpid_child.  However this only
works when a wait is in progress, which makes it impossible to
break out of a simple shell loop with ctrl-c:

   while true; do
   sleep 1
   echo hi
   done

Better performance can be achieved by setting the pending_int flag in
the control handler, similar to what's done in the signal handler.
The deferred signal is processed in int_on.

Or at least it would be but for a bug in the WIN32 code to handle
shellexec:  interrupts weren't being properly turned off and on.
So this had to be fixed too.