ash: allow long-running nofork applets to be interrupted
Nofork applets can't be interrupted with ctrl-c.  This isn't an
issue for most such applets because they do very little and won't
run for very long.  However 'yes' and 'seq 10000000' can't be
interrupted in a interactive shell, which is awkward.

As a special case ignore the nofork-ness of these applets if they're
run from an interactive shell.

This isn't foolproof as there are still ways to run them such that
they can't be interrupted, but it helps.