sed: support the -b/--binary option
This option was introduced in Cygwin's and MSYS2's sed, to allow for
keeping Carriage Returns (otherwise, they would be stripped and the
output would always be LF-only).

Git for Windows' test suite relies on the presence of this option.

It was easy enough to implement, too.

Conditional compilation added by rmy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Ron Yorston <rmy@pobox.com>