mingw: clarify that waitpid(-1, ...) is not supported
When passing -1 as pid to the waitpid() function, it is supposed to wait
for *any* child to exit. That is a bit tough to emulate on Windows given
that waitpid() returns only one pid, and if multiple children have exited,
subsequent calls to waitpid() should produce all of their pids
subsequently. Oh, and we would have to figure out which processes were
spawned and remember that list. And then, it would still be possible for
the children to exit in the meantime and for *another* process using the
pid, as Windows reuses pids very, very quickly.

So let's punt and simply state that we do not support that
functionality.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>