win32: make vfork fail at compile-time
Nothing in busybox-w32 uses fork.  Remove the implementations of fork,
vfork and xvfork so that, for example, enabling cpio pass-through causes
an error at compile-time, not run-time.

MinGW-w64 defines pid_t, so we don't have to.

Ensure code containing xvfork in netcat is disabled when NC_EXTRA
is disabled.  Otherwise removing the implementation of vfork costs
64 bytes.