win32: move detection of file formats to stat(2)
Move the code to detect shell scripts and binary executables from
mingw_access to a separate function, has_exec_format.  Call this function
in do_lstat to decide whether to set the executable bits in the file mode.
This will slow down stat but has a couple of advantages:

- shell scripts are highlighted in ls output
- the test applet can use stat(2) to detect executable files

The new function is used to handle another corner case in spawnveq:
binary executables without the usual .exe extension are only run by
spawnve if the file name ends with '.'.

Two minor changes:

- file_is_win32_executable has been renamed add_win32_extension to
  clarify what it does
- a call to file_is_executable has been removed from find_command
  in ash as it resulted in unhelpful error messages.