win32: additional improvements to handling of executables
Consistent processing of file extensions, as described in the previous
commit, has been applied to the 'which' applet and the functions
find_executable and mingw_spawn_interpreter.

In spawnveq check that the file to be executed exists and is executable,
and ensure that it won't have any extensions added by spawnve.

It's intended that all files passed to spawnve should have their names
fully specified.  If this isn't the case the tests here will cause errors
which will need to be fixed.