winansi: check for broken pipe
Microsoft Windows doesn't support SIGPIPE and its support for the
EPIPE error is somewhat inadequate.  Most of the time a broken pipe
leads to an EINVAL error which is misleading:

   $ yes hello | head -1
   hello
   yes: Invalid argument

Add a function to test for the underlying Windows error that indicates
a broken pipe and set errno to EPIPE.  Call this function in a few
strategic places.