ash: revise handling of builtin environment variables during fork
Commit db03ff5ea made some improvements to the handling of builtin
environment variables during forkshell.  However, more can be done.

Currently the calculated size of varinit doesn't match what's actually
copied.  It overestimates the amount of string space and the number
of pointers.  The text of each builtin environment variable is also
copied twice:  once as part of vartab and again from varinit.

Remove the code to copy varinit via the forkshell block.  Instead
use information from vartab and varinit_data to reinitialise varinit
and fix the links in vartab.

In addition, ensure that the LINENO variable is properly initialised
in the child.