Uppercase only the PATH env variable
busybox's env behaves differently, depending on how it is
called:

busybox.exe env  # env variable names casing is preserved.

or

busybox.exe sh -c "env"  # env variable names are uppercased.

But env on Linux preserves the casing, and so should busybox,
with the exception of the PATH variable, which must be uppercased,
otherwise the default value will be used instead, which can lead to
commands not being found.