ash: redefine SAVE_PTR macros to remove test
The SAVE_PTR macros are used to identify pointers that need to
be fixed after forkshell.  They're initially used in contexts
where the nodeptr variable may be NULL, so have a test for this
condition.

Later uses are in places where nodeptr is known to have a non-NULL
value so the macros can be redefined to remove the test.  Saves
over 100 bytes.