win32: special treatment for files with the '.sh' extension
Files with the extension '.sh' are considered to be executable.
Those that start with a '#!' line will be run using the specified
interpreter.  If no '#!' is present the script will be run by the
shell.

When searching for an executable the '.sh' extension will be
tested in the same way as the standard extensions understood
by spawnve().  '.sh' takes precedence over the standard
extensions.