ash: prevent ctrl-c from killing background processes
spawn_forkshell() was missing some code that would normally have
been run from forkchild when a background process was started:

- the ctrl-c handler wasn't disabled;
- standard input wasn't connected to /dev/null.

Using ctrl-c to kill a foreground process also killed background
processes.

Fixing this requires passing the forkshell mode and the number of
processes associated with the forkshell job to the child process.