win32: special treatment for PATH
The PATH shell variable is a special case.  It can be exported to
the environment where it might be interpreted by native applications
which assume the separator is ';'.  Hence:

- require that the separator used in PATH is ';'
- enforce this by intercepting calls to setvareq() that set PATH
  and adjusting its value if necessary.

As a result of this the code to parse PATH can be simplified by
replacing the hardcoded Unix ':' path separator by the platform-
dependent macro PATH_SEP.

The MANPATH variable is also required to use ';' as its separator
but since it's less likely to be used this isn't enforced.