ash: fix local PATH assignments
The previous commit included code to use PATH values assigned to
local variables:

   PATH=/new/path:$PATH new_binary

It didn't work in the case when no fork was needed, for example:

   (PATH=/new/path:$PATH new_binary)

because the call to listsetvar() freed the PATH value that was
passed to shellexec().  The forking case only worked because the
spawn_forkshell() mechanism in busybox-w32 didn't allow the PATH
value to be freed.

Rewrite the code so it works whether or not a fork is needed.  The
code would also work on POSIX systems.

A cleaner solution could be extracted from dash commit cbb71a8
'eval: Add assignment built-in support again' but neither that nor
the current patch are suitable for submission to upstream BusyBox.
This one is preferred because it requires fewer changes from
upstream.