busybox: add --uninstall option
Add an option to allow hard links to be removed.

   busybox --uninstall file

removes all hard links to the given file (including the file itself.)
Since Microsoft Windows refuses to delete a running executable a
BusyBox binary is unable to remove links to itself.

   busybox --uninstall -n file

displays the names of all hard links to the given file.

Although this feature is couched in terms of uninstalling BusyBox
it's actually quite general:  it can be used to delete or display
hard links to any file.