win32: implement umask(2)
umask() in the Microsoft C runtime takes different arguments to
umask(2).  Implement a fake umask(2) that remembers the mask and
calls the Microsoft umask() with an appropriate value.

Since the mask won't be inherited by children use an environment
variable to pass any value set by the shell built-in umask.