vi: display pasted text
vi tries to avoid updating the display if more input is available.
This didn't work when text was pasted in because a key release
event was left in the event queue.

When mysleep() is called to test for this condition allow up to one
event to be present in the queue before reporting that input is
available.