win32: allow characters to be entered as ALTNNN
It wasn't possible to enter characters using Alt and the decimal
character code.  This was because the character is generated when
the Alt key is released but the WIN32 implementation of read_key()
ignored all key up events.

Modify read_key() to ignore numbers entered on the numeric keypad
with Alt pressed and to detect when Alt is released.

Fixes GitHub issue #136.