ash: remove carriage returns from strings to be evaluated
The shell could fail to evaluate strings containing carriage
returns.  For example:

   awk 'BEGIN {
      "set -ex\r\npwd\r\n" | getline
   }' </dev/null

The string is passed as an argument to "sh -c".  The "set" built-in
fails because it attempts to treat the carriage return as an option.

Although this is correct behaviour on Unix it may be unhelpful on
Microsoft Windows.

See GitHub issue #138.