ash: make spawn_forkshell more like forkshell
These are cosmetic changes only.

Make the arguments to spawn_forkshell more like those of forkshell.
Since job control isn't currently supported in busybox-w32 the node
argument isn't actually used.

The node passed in the forkshell structure is used in the child
code.  It may be different from the node passed as an argument to
spawn_forkshell which, in the upstream code, is forwarded to
forkparent and forkchild.