win32: implement readlink(2)
Provide an implementation of readlink(2) based on code from Git
for Windows.  This version only supports symbolic links, not
mount points, as the latter seem to work well enough as-is.

With this change the ls and stat applets can display the targets
of symbolic links.  The readlink applet has been enabled in the
default configuration.