win32: canonicalize path in chdir(2)
Provide an implementation of chdir(2) which canonicalizes the
path to resolve symlinks.  Otherwise changing to a symlinked
directory confuses 'ls -l' because it thinks '.' is a link
rather than a directory.

OTOH, using 'cd' in the shell to change to a symlinked directory
now results in a mismatch between the shell's idea of where we are
and what's displayed in the prompt.  But upstream BusyBox does
that too so it must be OK.