su: add a basic implementation for WIN32
Use the undocumented 'runas' verb in a call to ShellExecuteEx()
to run a shell with elevated privileges.

Because of the way ShellExecuteEx() works this:

- requires that you acknowledge a User Account Control prompt (if
  you're an Administrator);
- requires that you enter an Administrator's password (if you aren't
  an Administrator);
- creates a separate console window for the privileged shell.

Variables from the parent shell aren't passed to its privileged child,
only variables from the environment.

It's possible to specify a command to run when the shell starts.
This can be used to pass shell variables:

  su -c "HELLO='hello world'; GOODBYE=$GOODBYE"

Or do fancy things like:

  su -c "ls -l; read -p 'Hit return to exit: '; exit"

It's probably best to put double quotes around the command and use
single quotes inside it.  Apparently ShellExecuteEx() requires
double quotes to be entered in triplicate:

  su -c 'HELLO="""hello world"""'