win32: extend normalisation of paths in realpath(3)
The code to normalise paths in resolve_symlinks(), which is used
by realpath(3), was incomplete and unable to handle UNC paths.
Make an ASCII version of normalize_ntpath() to extend the cases
covered.

This fixes a regression introduced by commit 585d17d26 (win32:
canonicalize path in chdir(2)):  it wasn't possible to change
to a directory with a UNC path.