lineedit: add a command to change backslashes to slashes
When a path name is copied to a console application using drag and
drop the path separator is backslash.  To handle this situation in
the shell add an editing command (Ctrl-Z) to convert all backslashes
on the current line to slashes.

See GitHub issue #149.

Also remove some unused code from read_key().