ash: set current working directory on startup
When the shell starts behave as though the user had manually changed
to the current working directory of the process, but only if it's an
interactive, non-login shell and the '-d' flag isn't present.

This gives more consistent behaviour when the shell is started in a
drive mapped to a network share or a symlinked directory.

This completes the changes requested in GitHub issue #148.