df: display origin of mapped drive
When a drive is mapped to a network share or a path display the
mapping in the 'Filesystem' column of df's output.

Since this changes the mnt_fsname field of the mntent structure
a slight alteration is needed to print_all_cwd() in ash.

Revert the change in commit a8c63f25b that enabled FEATURE_DF_FANCY
in the default configuration.  None of the additional options is
very useful.

See GitHub issue #164.