reset: enable in default configuration
The 'reset' applet attempts to reset the screen.  It's more thorough
than 'clear'.

Since it tries to reset the screen buffer for the current process
it has to be a NOFORK applet.

To run the Windows executable C:/Windows/System32/reset.exe use the
full path, as the 'reset' applet will take precedence.