ash: revert change to behaviour of -s option
Upstream BusyBox commit f3634584d (ash,hush: show 'c' in $- if run in
"sh -c CMD") changed the behaviour of -s to match that of bash rather
than dash.  This means that the -s and -c options can't be combined.

This combination of options is used by the Windows 'su' applet to
prevent the new console window from shutting down after the command
has run.  It's also useful when the shell is to be run with an
initialisation command (GitHub issue #100).

Revert the change so -s behaves as in dash.