winansi: fix escape processing in Windows 10 release 1809
As reported in GitHub issue #177, ANSI escape sequences don't work
reliably after shell redirection.

The problem appears to be due to a change in the behaviour of dup2()
in Windows 10 release 1809.  The escape handling code kept a copy of
the console handle for internal use.  Prior to Windows 10 release 1809
this worked; since then unwinding shell redirections results in the actual
console handle changing from time to time for reasons that are unknowable
without access to the source code.

Fix this by always fetching the console handle instead of caching it.