ash: rework how pointers are adjusted during forkshell
The original code kept an array of pointers to pointers that needed
to be adjusted after forkshell (nodeptr).  Switch to using a map to
indicate which locations in the data block need to be adjusted.

This requires a larger data block.  However the *_size() functions
which calculate the data block size need less information about the
data structures.  Saves about 100 bytes in the binary.