ash: reduce size of forkshell relocation map
The functions to calculate the size of the forkshell data block
returned the total size of funcblock and funcstring.  The relocation
map only needs to cover the size of the forkshell structure plus
funcblock, not funcstring.

Revise the *_size() functions to distinguish between funcblock and
funcstring.  Avoid complicating the calculation of node sizes
(calcsize() and related functions).  This slightly overestimates the
size of funcblock by including strings stored in node structures.

This change increases the binary by 96 bytes but can save several
thousand bytes at runtime.