mingw: fix off-by-one error in mingw_strftime()
When a format specification is replaced the loop variable 't'
should point to the last character of the replacement string in
the new format buffer.

In an extreme case if the original format string is "%z" and
tm->tm_isdst is negative to indicate that no DST information is
available the replacement string will be empty and 't' will point
to the location before the start of the new format buffer.  This
is OK.