winansi: fix alternate screen buffer in Windows 7
Closing 'vi' in Windows 7 resulted in the console window failing to
echo any output correctly until the 'reset' builtin was run.  This
didn't happen in Windows XP, 8 or 10.

The problem is fixed by only duplicating the console handle when switching
*to* the alternate screen buffer.

Add sanity checks so that switching to the alternate buffer requires
the cached handled to be invalid, while it must be valid when switching
from the alternate buffer.