ash: process backspace in read builtin
Previously the 'read' builtin had no special treatment for backspace
characters entered in interactive mode:  they were simply added to
the string being read.

Change this so that backspace deletes the previous character from
the buffer and updates the display to match.  It still doesn't
handle tabs in the input or lines that are larger than the console
width.