ash: catch EOF in read builtin
The read builtin didn't detect EOF so it wasn't possible to cleanly
terminate a loop like:

   while read -r; do echo $REPLY; done

Consider how Linux handles EOF (represented as ^D, though this isn't
echoed):

   $ awk '{print}' | xxd
   abc^D123
   ^D
   00000000: 6162 6331 3233 0a                        abc123.

Contrast with busybox-w32 on Windows (where the ^Z is echoed):

   $ awk '{print}' | xxd
   abc^Z123
   ^Z
   00000000: 6162 631a 3132 330a                      abc.123.

In both cases EOF is only detected at the start of a line.  On Linux
EOF within a line is dropped; on Windows it's output as a literal ctrl-Z.

Implement similar behaviour for the read builtin.