win32: changes to '#!' support
Since the earliest days of busybox-w32 '#!' has searched PATH for
the interpreter.  This doesn't seem to be supported by precedent.
In testing I also found that additional code was needed for the case
where PATH has been altered in the current shell or is modified on
the command line:  PATH has to be extracted from envp rather than
getenv("PATH").  Drop this non-standard feature.

*NIX implementations disagree on whether the interpreter can itself
be a script.  Follow Linux and allow this with a limit of four levels
of nesting.