ash: don't initialise shell variables in forkshell child
The forkshell child process calls init(), which mostly just initialises
shell variables.  Since these variables are overwritten with values from
the parent process there's no point in doing that.

In particular, init() initialises HOSTNAME by calling gethostname()
which involves network access and is slow.  Avoid this cost when
we know the shell is a forkshell child.