win32: use lazy loading for certain DLLs
Only a handful of functions are used from shell32.dll, userenv.dll
and psapi.dll.  Mostly these functions are in out of the way places.

By loading the functions only when required we can avoid the startup
cost of linking the three DLLs in the common case that they aren't
needed.