ash: minimise work when a forkshell process is started
There's no need to create or initialise the global data structures
when a forkshell is being started:  they'll soon be replaced by
data from the forkshell data block.

Move the call to forkshell_init() to the top of ash_main() along
with some initialisation that's common to forkshell shells and
ordinary shells.  Add the minimum necessary initialisation to
forkshell_init().