ash: treat all applets as NOEXEC
In standalone shell mode tryexec() treats all applets as NOEXEC.
Except for the shell:  calling ash_main() recursively seems unwise.

This avoids creating a new process for the applet (which is a Good
Thing on Microsoft Windows where creating processes is expensive) but
leaves the resources used by the invoking shell in the current process.