win32: enable globbing by default
Change how busybox.exe expands wildcards on the command line.

When globbing is enabled at compile time provide an implementation
of _setargv(), which is run early during startup of C programs.  This:

- enables globbing by setting _dowildcard to -1
- checks for the presence of the environment BB_GLOBBING
- if it exists and is set to 0 disables globbing
- if it doesn't exist sets BB_GLOBBING=0 but continues to apply
  Windows' globbing in the current process

The consequences of this are:

- When busybox.exe is initially run from a Command Prompt Windows'
  globbing is applied;
- Windows' globbing is turned off for future child processes, thus
  allowing the shell re-execute busybox.exe without it interfering
  with wildcards;
- this behaviour can be overridden by setting BB_GLOBBING explicitly.

Globbing can still be disabled at compile time if required.  In that
case BB_GLOBBING has no effect.

With these changes globbing can be enabled by default and BusyBox
will do the right thing in most circumstances.

(See GitHub issues #172 and #189.)