ash: changes to ctrl-c handling
Make ctrl-c handling more like Unix.  Remove the hSIGINT event and the
code in waitpid_child() to catch it.  Add a call to raise_interrupt()
in ctrl_handler().

Prior to these changes interrupts in a child shell weren't properly
handled.  In release FRP-3466-g53c09d0e1 interrupting a sleep in a
child shell results in both shells competing for input:

   ~ $ sh
   ~ $ sleep 10
   ^C^C

   ~ $ ~ $ pwd
   sh: w: not found
   ~ $
   sh: pd: not found
   ~ $

With recent changes:

   ~ $ sh
   ~ $ sleep 10
   ^C
   ~ $ echo $?
   130
   ~ $ exit
   ^C
   ~ $ echo $?
   130
   ~ $

The error return from the parent shell is due to the lack of job
control.  Upstream BusyBox ash and dash both do the same when job
control is disabled.