win32: use built-in applets for non-existent binaries with Unix-style paths
Shell scripts moved from Unix may contain hard-coded paths to
binaries such as /bin/sh.  A recent commit made it possible to
execute such binaries reliably, but that does require them to be
installed.  As an alternative solution:  if a binary with a
standard Unix path prefix can't be found but is available as a
built-in applet, run the applet.

Add the function unix_path() to detect paths starting with /bin,
/usr/bin, /sbin or /usr/sbin.

Use this function in:

- the 'which' applet
- shellexec(), describe_command() and find_command() in ash
- mingw_spawn_1()

See GitHub issue #195.