ls: allow backslashes to be replaced in displayed paths
Normally 'ls' displays paths exactly as the user enters them:

   $ ls .//file
   .//file

In busybox-w32 paths using backslash as a separator are displayed in
a form that can't be reused as input to the shell:

   $ ls .\\file
   .\file

Allow backslashes to be replaced with forward slashes if the
environment variable BB_FIX_BACKSLASH is set to 1.

See GitHub issue #196.