win32: allow putenv() to set empty values
WIN32 _putenv() can't set an environment variable with an empty
value because 'NAME=' is treated as a request to delete the variable.

Allow mingw_putenv() to set an empty value by first setting a fake
value and then truncating it.

This problem has always been present in mingw_putenv() but it became
particularly pressing when ash started treating all applets as NOEXEC.
The environment of NOEXEC applets is created by clearing the current
environment and using putenv() to set a new one from the shell
variables.  Empty variable were not being set.

See GitHub issue #197.