build system: use static ssp library
The previous commit got rid of a link error but generates a binary
that isn't standalone:  it requires libssp-0.dll.

Force the use of the static ssp library.  This also requires mingwex
to be added to LDLIBS, since that's what has a dependency on ssp.

Other workarounds are to compile with non-zero _FORTIFY_SOURCE (which
increases the size of the binary) or to add '-u ___strcpy_chk' to
LDFLAGS (likely to be fragile).